Add more re-organization of modules for xgboost model
